moved JSON read and write to separate file
