+++ baseclasses/utils/__init__.py
@@ -1,5 +1,13 @@
 from .containers import CaseInsensitiveSet, CaseInsensitiveDict
 from .error import Error
 from .utils import getPy3SafeString
+from .fileIO import writeJSON, readJSON
 
-__all__ = ["CaseInsensitiveSet", "CaseInsensitiveDict", "Error", "getPy3SafeString"]
+__all__ = [
+    "CaseInsensitiveSet",
+    "CaseInsensitiveDict",
+    "Error",
+    "getPy3SafeString",
+    "writeJSON",
+    "readJSON",
+]
